requirements.txt
apache-airflow==2.8.1
beautifulsoup4==4.11.1
selenium==4.18.1
lxml==5.1.0
requests==2.32.5
yfinance==1.0
pandas==1.5.3
numpy==1.26.4
openpyxl==3.1.0
jellyfish==1.2.1
transformers==4.44.0
torch==2.5.1
tokenizers==0.19.1
safetensors==0.5.3
sentencepiece==0.2.1
scipy==1.11.4
scikit-learn>=1.3.0
fredapi>=0.5.0
